support column debugger

+++ ecmascript/tooling/base/pt_events.cpp
@@ -243,7 +243,7 @@ std::unique_ptr<ScriptFailedToParse> ScriptFailedToParse::Create(const EcmaVM *e
     result = Local<ObjectRef>(params)->Get(ecmaVm, Local<JSValueRef>(StringRef::NewFromUtf8(ecmaVm, "startLine")));
     if (!result.IsEmpty() && !result->IsUndefined()) {
         if (result->IsNumber()) {
-            scriptEvent->startLine_ = static_cast<int32_t>(Local<NumberRef>(result)->Value());
+            scriptEvent->startLine_ = static_cast<size_t>(Local<NumberRef>(result)->Value());
         } else {
             error += "'startLine' should a Number;";
         }
@@ -253,7 +253,7 @@ std::unique_ptr<ScriptFailedToParse> ScriptFailedToParse::Create(const EcmaVM *e
     result = Local<ObjectRef>(params)->Get(ecmaVm, Local<JSValueRef>(StringRef::NewFromUtf8(ecmaVm, "startColumn")));
     if (!result.IsEmpty() && !result->IsUndefined()) {
         if (result->IsNumber()) {
-            scriptEvent->startColumn_ = static_cast<int32_t>(Local<NumberRef>(result)->Value());
+            scriptEvent->startColumn_ = static_cast<size_t>(Local<NumberRef>(result)->Value());
         } else {
             error += "'startColumn' should a Number;";
         }
@@ -263,7 +263,7 @@ std::unique_ptr<ScriptFailedToParse> ScriptFailedToParse::Create(const EcmaVM *e
     result = Local<ObjectRef>(params)->Get(ecmaVm, Local<JSValueRef>(StringRef::NewFromUtf8(ecmaVm, "endLine")));
     if (!result.IsEmpty() && !result->IsUndefined()) {
         if (result->IsNumber()) {
-            scriptEvent->endLine_ = static_cast<int32_t>(Local<NumberRef>(result)->Value());
+            scriptEvent->endLine_ = static_cast<size_t>(Local<NumberRef>(result)->Value());
         } else {
             error += "'endLine' should a Number;";
         }
@@ -273,7 +273,7 @@ std::unique_ptr<ScriptFailedToParse> ScriptFailedToParse::Create(const EcmaVM *e
     result = Local<ObjectRef>(params)->Get(ecmaVm, Local<JSValueRef>(StringRef::NewFromUtf8(ecmaVm, "endColumn")));
     if (!result.IsEmpty() && !result->IsUndefined()) {
         if (result->IsNumber()) {
-            scriptEvent->endColumn_ = static_cast<int32_t>(Local<NumberRef>(result)->Value());
+            scriptEvent->endColumn_ = static_cast<size_t>(Local<NumberRef>(result)->Value());
         } else {
             error += "'endColumn' should a Number;";
         }
@@ -337,7 +337,7 @@ std::unique_ptr<ScriptFailedToParse> ScriptFailedToParse::Create(const EcmaVM *e
     result = Local<ObjectRef>(params)->Get(ecmaVm, Local<JSValueRef>(StringRef::NewFromUtf8(ecmaVm, "length")));
     if (!result.IsEmpty() && !result->IsUndefined()) {
         if (result->IsNumber()) {
-            scriptEvent->length_ = static_cast<int32_t>(Local<NumberRef>(result)->Value());
+            scriptEvent->length_ = static_cast<uint32_t>(Local<NumberRef>(result)->Value());
         } else {
             error += "'length' should a Number;";
         }
@@ -345,7 +345,7 @@ std::unique_ptr<ScriptFailedToParse> ScriptFailedToParse::Create(const EcmaVM *e
     result = Local<ObjectRef>(params)->Get(ecmaVm, Local<JSValueRef>(StringRef::NewFromUtf8(ecmaVm, "codeOffset")));
     if (!result.IsEmpty() && !result->IsUndefined()) {
         if (result->IsNumber()) {
-            scriptEvent->codeOffset_ = static_cast<int32_t>(Local<NumberRef>(result)->Value());
+            scriptEvent->codeOffset_ = static_cast<uint32_t>(Local<NumberRef>(result)->Value());
         } else {
             error += "'codeOffset' should a Number;";
         }
@@ -496,7 +496,7 @@ std::unique_ptr<ScriptParsed> ScriptParsed::Create(const EcmaVM *ecmaVm, const L
     result = Local<ObjectRef>(params)->Get(ecmaVm, Local<JSValueRef>(StringRef::NewFromUtf8(ecmaVm, "startLine")));
     if (!result.IsEmpty() && !result->IsUndefined()) {
         if (result->IsNumber()) {
-            scriptEvent->startLine_ = static_cast<int32_t>(Local<NumberRef>(result)->Value());
+            scriptEvent->startLine_ = static_cast<size_t>(Local<NumberRef>(result)->Value());
         } else {
             error += "'startLine' should a Number;";
         }
@@ -506,7 +506,7 @@ std::unique_ptr<ScriptParsed> ScriptParsed::Create(const EcmaVM *ecmaVm, const L
     result = Local<ObjectRef>(params)->Get(ecmaVm, Local<JSValueRef>(StringRef::NewFromUtf8(ecmaVm, "startColumn")));
     if (!result.IsEmpty() && !result->IsUndefined()) {
         if (result->IsNumber()) {
-            scriptEvent->startColumn_ = static_cast<int32_t>(Local<NumberRef>(result)->Value());
+            scriptEvent->startColumn_ = static_cast<size_t>(Local<NumberRef>(result)->Value());
         } else {
             error += "'startColumn' should a Number;";
         }
@@ -516,7 +516,7 @@ std::unique_ptr<ScriptParsed> ScriptParsed::Create(const EcmaVM *ecmaVm, const L
     result = Local<ObjectRef>(params)->Get(ecmaVm, Local<JSValueRef>(StringRef::NewFromUtf8(ecmaVm, "endLine")));
     if (!result.IsEmpty() && !result->IsUndefined()) {
         if (result->IsNumber()) {
-            scriptEvent->endLine_ = static_cast<int32_t>(Local<NumberRef>(result)->Value());
+            scriptEvent->endLine_ = static_cast<size_t>(Local<NumberRef>(result)->Value());
         } else {
             error += "'endLine' should a Number;";
         }
@@ -526,7 +526,7 @@ std::unique_ptr<ScriptParsed> ScriptParsed::Create(const EcmaVM *ecmaVm, const L
     result = Local<ObjectRef>(params)->Get(ecmaVm, Local<JSValueRef>(StringRef::NewFromUtf8(ecmaVm, "endColumn")));
     if (!result.IsEmpty() && !result->IsUndefined()) {
         if (result->IsNumber()) {
-            scriptEvent->endColumn_ = static_cast<int32_t>(Local<NumberRef>(result)->Value());
+            scriptEvent->endColumn_ = static_cast<size_t>(Local<NumberRef>(result)->Value());
         } else {
             error += "'endColumn' should a Number;";
         }
@@ -598,7 +598,7 @@ std::unique_ptr<ScriptParsed> ScriptParsed::Create(const EcmaVM *ecmaVm, const L
     result = Local<ObjectRef>(params)->Get(ecmaVm, Local<JSValueRef>(StringRef::NewFromUtf8(ecmaVm, "length")));
     if (!result.IsEmpty() && !result->IsUndefined()) {
         if (result->IsNumber()) {
-            scriptEvent->length_ = static_cast<int32_t>(Local<NumberRef>(result)->Value());
+            scriptEvent->length_ = static_cast<uint32_t>(Local<NumberRef>(result)->Value());
         } else {
             error += "'length' should a Number;";
         }
@@ -606,7 +606,7 @@ std::unique_ptr<ScriptParsed> ScriptParsed::Create(const EcmaVM *ecmaVm, const L
     result = Local<ObjectRef>(params)->Get(ecmaVm, Local<JSValueRef>(StringRef::NewFromUtf8(ecmaVm, "codeOffset")));
     if (!result.IsEmpty() && !result->IsUndefined()) {
         if (result->IsNumber()) {
-            scriptEvent->codeOffset_ = static_cast<int32_t>(Local<NumberRef>(result)->Value());
+            scriptEvent->codeOffset_ = static_cast<uint32_t>(Local<NumberRef>(result)->Value());
         } else {
             error += "'codeOffset' should a Number;";
         }
